Separate repo profile panel sections to separated components
